Del helpers.js file: move constants to /services/const/, move functions to /services/utils
