+++ src/services/utils/castTimeFormat.js
@@ -1,0 +1,3 @@
+export default (value) => {
+  return value < 10 ? `0${value}` : String(value);
+};
