+++ src/services/utils/getCorrectTime.js
@@ -1,12 +1,7 @@
-import {MONTHS} from "./const";
+import MONTHS from "../const/months";
+import castTimeFormat from "./castTimeFormat";
 
-export const getRandomInt = (min, max) => min + Math.floor(Math.random() * (max + 1 - min));
-
-export const castTimeFormat = (value) => {
-  return value < 10 ? `0${value}` : String(value);
-};
-
-export const getCorrectTime = (date) => {
+export default (date) => {
   const day = castTimeFormat(date.getDate());
   const monthIndex = date.getMonth();
   const month = castTimeFormat(monthIndex + 1);
@@ -23,8 +18,4 @@ export const getCorrectTime = (date) => {
     string: `${day}/${month}/${year} ${hours}:${minutes}`,
     stringISO: `${year}-${month}-${day}T${hours}:${minutes}`
   };
-};
-
-export const render = (container, template, place = `beforeend`) => {
-  container.insertAdjacentHTML(place, template);
 };
