+++ src/services/utils/render.js
@@ -1,0 +1,3 @@
+export default (container, template, place = `beforeend`) => {
+  container.insertAdjacentHTML(place, template);
+};
